Fix: remove now verbose notification body and related comment
I learned that the actual execution times set by setTimeout are not so much delayed.
I have speculated that I should shorten the time to setTimeout, but found it's not necessary.

So I delete the commented-out constant to calcuate the shortened time
and notification body to see how long the execution times are delayed.