Fix error when all the saved times have been done
When (re)loading the web service after filling the form with all passed
times, the worker complains that it can't get millisecsAfter method of
`undefined`.
It's because `this.nextTimeIndex_` is `-1` when all the input times have
been passed.
(i.e. `this.times_[this.nextTimeIndex_].millisecsAfter > 0` returns false)

So I rewrote the algorithm to calculate next time to run the task in more
functional-programmer-preferrable-manner to avoid null reference.